MCP_AGENTS/requirements.txt
google-genai>=1.14.0
exa-py>=1.5.0
pydantic>=2.7.0
python-dotenv>=1.0.0
mcp>=1.15.0
pypdf>=5.0.0
requests>=2.30.0
